pyproject.toml
[project]
name = "pioneer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "fal-client>=0.9.1",
    "langchain[mistralai,openai]>=0.3.27",
    "mistralai>=1.9.11",
    "openai>=2.8.0",
    "pydantic>=2.12.4",
    "python-decouple>=3.8",
    "redis>=7.0.1",
    "requests>=2.32.5",
]
